Initial V2.0 Upload
Initial commit of all the version 2.0 features that I have been working on in the background. Includes better detection methods, improved refund processing, better error catching and handling, statistic tracking, and so on.
Next ideas include removing remnants of base V1.* code and improving on it. Possibly removing the timeout sections entirely or minimizing it as it is bloating the code and replacing it by detecting refund success messages instead assuming I can figure out the issue of the detection not always working.